rename file base-http to http so its compatible with rails
