+++ examples/client.rb
@@ -1,6 +1,6 @@
 require_relative '../lib/grenache-ruby-http.rb'
 
-Grenache::Base.configure do |conf|
+Grenache::Http.configure do |conf|
    conf.grape_address = "http://127.0.0.1:40002/"
 end
